+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/LegacySnapshotConfiguration.java
@@ -1,19 +1,18 @@
 package de.skuzzle.test.snapshots.impl;
 
-import java.lang.reflect.Method;
-import java.nio.file.Path;
-
+import de.skuzzle.test.snapshots.EnableSnapshotTests;
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
 
-import de.skuzzle.test.snapshots.EnableSnapshotTests;
+import java.lang.reflect.Method;
+import java.nio.file.Path;
+import java.util.List;
 
 /**
- *
  * @author Simon Taddiken
  * @since 1.7.0
  * @deprecated Since 1.7.0 - Only introduced for compatibility reasons. Will be removed
- *             with 2.0.0
+ * with 2.0.0
  */
 @Deprecated(since = "1.7.0", forRemoval = true)
 @API(status = Status.DEPRECATED, since = "1.7.0")