+++ readme/RELEASE_NOTES.md
@@ -14,6 +14,7 @@
 * [#61](https://github.com/skuzzle/snapshot-tests/issues/61): `@SnapshotDirectory` is not allowed anymore on test methods, only on test class
 * [#64](https://github.com/skuzzle/snapshot-tests/issues/64): Migrate build to gradle
 * [#66](https://github.com/skuzzle/snapshot-tests/issues/66): XmlSnapshot.withEnableXPathDebugging: Boolean parameter 'enabled' not taken into account
+* [#71](https://github.com/skuzzle/snapshot-tests/issues/71): Gracefully support JUnit5 nested tests
 * Add recursive directory scanning via `DirectoriesFrom.recursive()`
 * Deprecate `PathFilter` in favor of `TestFileFilter` and `TestDirectoryFilter`
 * Improve format and information density of XPath debugging (can be enabled with `XmlSnapshot.withEnableXPathDebugging(true)`)